# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Project Mimir API
  description: |
    Project Mimir — AI-Powered Knowledge Management Platform API.

    ## Authentication
    Most endpoints require Bearer token authentication. Obtain a token via `POST /api/v1/auth/login`.

    ## Base URL
    - Development: `http://localhost:3000`
    - Production: Configured via environment
  version: 1.0.0
  contact:
    name: Project Mimir Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/sources/{id}:
    get:
      summary: Get source by ID
      tags: [Sources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Source details
    delete:
      summary: Delete source
      tags: [Sources]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Source deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...